Move ftml FFI classes to a subdirectory.
Avoid causing confusion with general wikitext service classes.
