+++ web/app/Services/Wikitext/FFI/TextOutput.php
@@ -1,13 +1,13 @@
 <?php
 declare(strict_types = 1);
 
-namespace Wikijump\Services\Wikitext;
+namespace Wikijump\Services\Wikitext\FFI;
 
 use \FFI;
 
 /**
  * Class TextOutput, representing a returned 'struct ftml_text_output' object.
- * @package Wikijump\Services\Wikitext
+ * @package Wikijump\Services\Wikitext\FFI
  */
 class TextOutput
 {
